requirements.txt
PySide6>=6.7
numpy>=1.26
pyaudio>=0.2.13
Pillow>=10.0
mido>=1.2.10
pyfluidsynth>=1.3.3
